music-emotion-recommender/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2 
spotipy==2.23.0
python-dotenv==1.0.0
pydantic==2.5.2